rename src/business_logic/fact_state/state_reader.rs -> src/business_logic/fact_state/in_memory_state_reader.rs (#84)
