+++ src/business_logic/fact_state/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod contract_state;
+pub(crate) mod in_memory_state_reader;
 pub mod state;
-pub(crate) mod state_reader;
